move config file to runner package
